building img update and deletion

Closes #42

See merge request web-inquilino-perfecto/web_project_inquilinoperfecto!84

+++ controllers/propertyController.js
@@ -48,7 +48,7 @@ const createNewProperty = async(req, res) =>{
 
         const createdProp = await save(newProp, tName)
 
-        const prevDir = propDirectory + '/' + req.auth.user.user_uuid
+        const prevDir = propDirectory + '/' + newProp.usr_casero_uuid
         const newDir = propDirectory + '/' + newProp.inmueble_uuid
         fs.renameSync(prevDir, newDir)
 
@@ -199,7 +199,6 @@ const getPropertiesSelf = async(req, res) =>{
     let isStatus, sendMessage;
     const tName = 'inmuebles';
     try {
-        console.log( req.auth.user.user_uuid);
         const propCasero = { usr_casero_uuid : req.auth.user.user_uuid }
         const selfProp = await findItems(propCasero,tName)
 
@@ -241,9 +240,11 @@ const getPropertiesSelf = async(req, res) =>{
 const modifyProperty = async(req, res) =>{
     let isStatus, sendMessage;
     const tName = 'inmuebles';
+    const tImgs = 'img_inmuebles';
     try {
         const oldProp = validateUuid(req.params)
         let existsProp = await findItems(oldProp, tName)
+        existsProp = existsProp[0]
         if(!existsProp){
             throw new errorNoEntryFound(
                 'Prop update by admin or self',
@@ -253,16 +254,34 @@ const modifyProperty = async(req, res) =>{
             )
         }
         else if(
-            existsProp.length > 0
-            &&(
-                req.auth?.user?.user_uuid === existsProp[0].usr_casero_uuid ||
-                req.auth?.user?.tipo === 'ADMIN')
+                req.auth?.user?.user_uuid === existsProp.usr_casero_uuid ||
+                req.auth?.user?.tipo === 'ADMIN'
         ){
-            console.log(req.body);
             let newProp = propUpdateValidate(req.body)
             newProp = {...oldProp, ...newProp}
             const consulta = await updateItem(newProp, oldProp, tName)
             if(consulta >= 1){
+
+                if(req.files){
+                    const prevDir = propDirectory + '/' + existsProp.usr_casero_uuid
+                    const newDir = propDirectory + '/' + newProp.inmueble_uuid
+                    if(fs.existsSync(prevDir+'/')){
+                        fs.rmdirSync(newDir, { recursive: true });
+                        const delPropImg = deleteItem({inmueble_uuid: newProp.inmueble_uuid},tImgs)
+                        fs.renameSync(prevDir, newDir)
+
+                        const filenames = fs.readdirSync(newDir)
+                        for(const f in filenames){
+                            const tuple = {
+                                img_inmueble_uuid: v4(),
+                                inmueble_uuid: newProp.inmueble_uuid,
+                                img_inmueble: newDir + '/' + filenames[f]
+                            }
+                            const saveIt = await save(tuple,tImgs)
+                        }
+                    }
+                }
+
                 isStatus = 200
                 sendMessage = {
                     info: "Inmueble modificado",
@@ -299,24 +318,34 @@ const modifyProperty = async(req, res) =>{
 const deleteProperty = async(req, res) =>{
     let isStatus, sendMessage;
     const tName = 'inmuebles';
+    const tImgs = 'img_inmuebles';
     try {
         const validatedDelProp = validateUuid(req.body)
         let existsProp = await findItems(validatedDelProp,tName)
-        existsProp = existsProp[0]
-        if(existsProp){
+
+        if(existsProp && existsProp.length!== 0){
+            existsProp = existsProp[0]
             if(
-                req.auth?.user?.user_uuid === existsProp.usr_casero_uuid 
+                req.auth?.user?.user_uuid === existsProp.usr_casero_uuid
                 ||req.auth?.user?.tipo === 'ADMIN'
             ){
                 const isPropDel = await deleteItem(validatedDelProp, tName)
-                sendMessage = {
-                    "tuple": validatedDelProp,
-                    "delete": isPropDel
-                }
-                isStatus = 200
-                console.log(isPropDel ?
+                if(isPropDel){
+                    const imgDir = propDirectory + '/' + validatedDelProp.inmueble_uuid
+                    if(fs.existsSync(imgDir)){
+                        fs.rmdirSync(imgDir, {recursive : true})
+                    }
+                    const isPropImgDel = await deleteItem(validatedDelProp, tImgs)
+
+                    sendMessage = {
+                        "tuple": validatedDelProp,
+                        "delete": isPropDel
+                    }
+                    isStatus = 200
+                    console.log(isPropDel ?
                         `Successfully deletion for ${Object.keys(validatedDelProp)[0]} with ${validatedDelProp.inmueble_uuid}`
                         : `No tuple could be deleted for ${Object.keys(validatedDelProp)[0]} with ${validatedDelProp.inmueble_uuid}`);
+                }
             }else{
                 throw new errorNoAuthorization(
                     req.auth?.user?.username,
